UIQM-255 quickMARC : Allow a user to tab from one subfield to another using a shortcut key in a single row
