+++ src/common/constants/keyboardCommands.js
@@ -1,10 +1,15 @@
+export const KEYBOARD_COMMAND_NAMES = {
+  NEXT_SUBFIELD: 'NEXT_SUBFIELD',
+  PREV_SUBFIELD: 'PREV_SUBFIELD',
+};
+
 export const keyboardCommands = [
   {
-    name: 'nextsubfield',
+    name: KEYBOARD_COMMAND_NAMES.NEXT_SUBFIELD,
     shortcut: 'ctrl+]',
   },
   {
-    name: 'prevsubfield',
+    name: KEYBOARD_COMMAND_NAMES.PREV_SUBFIELD,
     shortcut: 'ctrl+[',
   },
 ];
